+++ src/main/java/kuit/project/beering/dto/common/FieldValidationError.java
@@ -1,4 +1,4 @@
-package kuit.project.beering.util;
+package kuit.project.beering.dto.common;
 
 import lombok.*;
